backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
sqlalchemy==2.0.25
pydantic>=2.5.3
python-dotenv==1.0.0
httpx>=0.28.0
pypdf==3.17.4
pillow==10.2.0
supabase==2.28.0
pgvector==0.2.4
psycopg2-binary==2.9.9
